@@ -13,29 +13,28 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
     "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
     "@angular/pwa": "^0.13.9",
-    "@angular/router": "^7.2.15",
+    "@angular/router": "^8.2.14",
+    "@angular/service-worker": "^8.2.14",
     "bootstrap": "^4.3.1",
     "core-js": "^3.1.2",
     "jquery": "^3.4.1",
-    "ng5-breadcrumb": "0.0.6",
-    "rxjs": "^6.0.0",
-    "zone.js": "^0.9.1",
-    "@angular/service-worker": "^7.2.15"
+    "rxjs": "^6.5.5",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.9",
-    "@angular/cli": "~7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.26",
+    "@angular/cli": "~8.3.26",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/jasmine": "~3.3.12",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~12.0.2",
@@ -48,9 +47,10 @@
     "karma-coverage-istanbul-reporter": "~2.0.5",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
+    "node-sass": "^4.13.1",
     "protractor": "~5.4.2",
     "ts-node": "~8.1.1",
     "tslint": "~5.16.0",
-    "typescript": "~3.2.2"
+    "typescript": "~3.5.3"
   }
 }
