@@ -1,9 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 
-// Exteral Libs
-import { BreadcrumbService } from 'ng5-breadcrumb';
-
 @Component({
   selector: 'mc-about',
   templateUrl: './about.component.html',
@@ -11,14 +8,10 @@ import { BreadcrumbService } from 'ng5-breadcrumb';
 })
 export class AboutComponent implements OnInit {
 
-  constructor(
-    private breadcrumbService: BreadcrumbService,
-    private titleService: Title
-  ) { }
+  constructor(private titleService: Title) { }
 
   ngOnInit() {
     this.titleService.setTitle('Sobre');
-    this.breadcrumbService.addFriendlyNameForRouteRegex('/about', 'Sobre');
   }
 
 }
